Move final files to the correct folder
