+++ client/src/schema/index.ts
@@ -1,1 +1,3 @@
 export * from './operators';
+export type { XataRecord } from './record';
+export { Repository, RestRepository } from './repository';
